requirements.txt
python-pptx==0.6.21
pandas==1.4.2